fixed bug in modify patient + add delete all func
